feat: VisionArea 커스텀훅, 유틸함수 분리 및 타입 정의
