+++ src/featured/interview/components/VideoArea.tsx
@@ -4,59 +4,8 @@ import { motion, AnimatePresence } from 'framer-motion'
 import { Mic, MicOff, VideoOff, Activity, AlertTriangle, CheckCircle } from 'lucide-react'
 import type { Phase } from '@/featured/interview/types'
 import Webcam from 'react-webcam'
-import { useGazeTracker } from '../hooks/useGazeTracker'
-import { useEffect, useRef, useState } from 'react'
-
-// --- 1. 비전 연산 유틸리티 함수 ---
-
-const getDistance = (p1: { x: number; y: number }, p2: { x: number; y: number }) => {
-  return Math.hypot(p1.x - p2.x, p1.y - p2.y)
-}
-
-// scaleX/scaleY로 정규화 좌표 → 픽셀 변환 후 EAR 계산
-// (가로/세로 스케일이 다른 영상에서 aspect ratio 왜곡 방지)
-const calculateEAR = (landmarks: any[], indices: number[], scaleX: number, scaleY: number) => {
-  const pts = indices.map((i) => ({
-    x: landmarks[i].x * scaleX,
-    y: landmarks[i].y * scaleY,
-  }))
-  const v1 = getDistance(pts[1], pts[5])
-  const v2 = getDistance(pts[2], pts[4])
-  const h = getDistance(pts[0], pts[3])
-  if (h === 0) return 1
-  return (v1 + v2) / (2.0 * h)
-}
-
-// 💡 [수정됨] Matrix 파싱 안정성 강화
-const extractEulerAngles = (matrix: number[] | Float32Array) => {
-  try {
-    const RAD_TO_DEG = 180 / Math.PI
-    // MediaPipe의 4x4 변환 행렬에서 회전 정보 추출
-    const pitch = Math.atan2(matrix[6], matrix[10]) * RAD_TO_DEG // 상하 반전 보정
-    const yaw = -Math.asin(matrix[2]) * RAD_TO_DEG // 좌우 반전 보정
-
-    // NaN 방지 처리
-    return {
-      pitch: isNaN(pitch) ? 0 : pitch,
-      yaw: isNaN(yaw) ? 0 : yaw,
-    }
-  } catch (e) {
-    return { pitch: 0, yaw: 0 }
-  }
-}
-
-const LEFT_EYE = [33, 160, 158, 133, 153, 144]
-const RIGHT_EYE = [362, 385, 387, 263, 373, 380]
-const EAR_THRESHOLD = 0.22
-
-interface BackendPayload {
-  timestamp: string
-  focusState: 'CENTER' | 'LEFT' | 'RIGHT' | 'UP' | 'DOWN'
-  blinkCount: number
-  pitch: string
-  yaw: string
-  eventType: 'ANOMALY' | 'RECOVERY' // 백엔드 분류용 태그 추가
-}
+import { useVisionAnalysis } from '@/featured/interview/hooks/useVisionAnalysis'
+import { EAR_THRESHOLD } from '@/featured/interview/utils/visionUtils'
 
 interface VideoAreaProps {
   cameraOn: boolean
@@ -66,158 +15,14 @@ interface VideoAreaProps {
 }
 
 export function VideoArea({ cameraOn, micOn, phase, basePose }: VideoAreaProps) {
-  const landmarker = useGazeTracker()
-  const webcamRef = useRef<Webcam>(null)
-  const requestRef = useRef<number | null>(null)
-
-  const [realtimeStats, setRealtimeStats] = useState({
-    pitch: 0,
-    yaw: 0,
-    ear: 0,
-    focusState: 'CENTER',
+  const { backendLogs, realtimeStats, webcamRef, landmarker } = useVisionAnalysis({
+    cameraOn,
+    phase,
+    basePose,
   })
-
-  const [backendLogs, setBackendLogs] = useState<BackendPayload[]>([])
-
-  const blinkCountRef = useRef(0)
-  const isBlinkingRef = useRef(false)
-  const lastDetectTimeRef = useRef(0)
-  const lastStateUpdateRef = useRef(0)
-  const lastLogTimeRef = useRef(0)
-
-  // 💡 [추가됨] 이전 상태를 기억하여 상태가 변할 때만 로그 전송
-  const lastLoggedStateRef = useRef<'CENTER' | 'LEFT' | 'RIGHT' | 'UP' | 'DOWN'>('CENTER')
-
-  const DETECT_INTERVAL_MS = 50
-  const STATE_INTERVAL_MS = 100
-
-  useEffect(() => {
-    const detectGaze = () => {
-      requestRef.current = requestAnimationFrame(detectGaze)
-
-      const now = performance.now()
-      if (now - lastDetectTimeRef.current < DETECT_INTERVAL_MS) return
-      lastDetectTimeRef.current = now
-
-      const video = webcamRef.current?.video
-      if (!landmarker || !video || video.readyState !== 4) return
-
-      try {
-        const result = landmarker.detectForVideo(video, now)
-        if (!result.faceLandmarks?.length) return
-
-        const landmarks = result.faceLandmarks[0]
-
-        // 1. EAR (눈 깜빡임) — 픽셀 좌표 기준으로 계산
-        const scaleX = video.videoWidth
-        const scaleY = video.videoHeight
-        const avgEAR =
-          (calculateEAR(landmarks, LEFT_EYE, scaleX, scaleY) +
-            calculateEAR(landmarks, RIGHT_EYE, scaleX, scaleY)) /
-          2
-        if (avgEAR < EAR_THRESHOLD) {
-          if (!isBlinkingRef.current) {
-            blinkCountRef.current += 1
-            isBlinkingRef.current = true
-          }
-        } else {
-          isBlinkingRef.current = false
-        }
-
-        // 2. Head Pose (고개 각도)
-        let pitch = 0,
-          yaw = 0
-        const matrixes = result.facialTransformationMatrixes
-        if (matrixes && matrixes.length > 0) {
-          // MediaPipe 버전에 따라 data 객체 안에 있거나 배열 자체일 수 있음
-          const rawMatrix = matrixes[0].data || matrixes[0]
-          const angles = extractEulerAngles(rawMatrix)
-          pitch = angles.pitch
-          yaw = angles.yaw
-        }
-
-        // 3. 시선 집중도 (basePose 기준으로 상대 판단)
-        const baseYaw = basePose?.yaw ?? 0
-        const basePitch = basePose?.pitch ?? 0
-        let currentFocus: 'CENTER' | 'LEFT' | 'RIGHT' | 'UP' | 'DOWN' = 'CENTER'
-        if (yaw - baseYaw > 15) currentFocus = 'LEFT'
-        else if (yaw - baseYaw < -15) currentFocus = 'RIGHT'
-        else if (pitch - basePitch > 15) currentFocus = 'DOWN'
-        else if (pitch - basePitch < -10) currentFocus = 'UP'
-
-        // 4. UI 상태 업데이트
-        if (now - lastStateUpdateRef.current >= STATE_INTERVAL_MS) {
-          lastStateUpdateRef.current = now
-          setRealtimeStats({
-            pitch: Math.round(pitch),
-            yaw: Math.round(yaw),
-            ear: Number(avgEAR.toFixed(2)),
-            focusState: currentFocus,
-          })
-        }
-
-        // 💡 5. [수정됨] 특이점 발생 시에만 백엔드 전송 (Event-driven)
-        const isAnomaly = currentFocus !== 'CENTER'
-        const isStateChanged = currentFocus !== lastLoggedStateRef.current
-        const timeSinceLastLog = now - lastLogTimeRef.current
-
-        // 조건 1: 시선 이탈이 감지되었고 (상태가 막 변했거나, 이탈한 지 2초가 지났을 때)
-        if (isAnomaly && (isStateChanged || timeSinceLastLog > 2000)) {
-          lastLogTimeRef.current = now
-          lastLoggedStateRef.current = currentFocus
-
-          setBackendLogs((prev) =>
-            [
-              {
-                timestamp: new Date().toISOString().split('T')[1].slice(0, 8),
-                focusState: currentFocus,
-                blinkCount: blinkCountRef.current,
-                pitch: pitch.toFixed(1),
-                yaw: yaw.toFixed(1),
-                eventType: 'ANOMALY' as const,
-              },
-              ...prev,
-            ].slice(0, 5),
-          )
-          blinkCountRef.current = 0
-        }
-        // 조건 2: 시선이 이탈했다가 다시 정면('CENTER')으로 복귀했을 때 (1회성 기록)
-        else if (!isAnomaly && isStateChanged) {
-          lastLogTimeRef.current = now
-          lastLoggedStateRef.current = 'CENTER'
-
-          setBackendLogs((prev) =>
-            [
-              {
-                timestamp: new Date().toISOString().split('T')[1].slice(0, 8),
-                focusState: 'CENTER' as const,
-                blinkCount: blinkCountRef.current,
-                pitch: pitch.toFixed(1),
-                yaw: yaw.toFixed(1),
-                eventType: 'RECOVERY' as const,
-              },
-              ...prev,
-            ].slice(0, 5),
-          )
-          blinkCountRef.current = 0
-        }
-      } catch (error) {
-        console.error('MediaPipe Error:', error)
-      }
-    }
-
-    if (landmarker && cameraOn && phase === 'answering') {
-      requestRef.current = requestAnimationFrame(detectGaze)
-    }
-
-    return () => {
-      if (requestRef.current) cancelAnimationFrame(requestRef.current)
-    }
-  }, [landmarker, cameraOn, phase])
 
   return (
     <div className="flex w-full max-w-175 flex-col gap-4">
-      {/* 윗부분: 기존 비디오 영역 (변경 없음) */}
       <motion.div
         layout
         className="relative flex aspect-video w-full items-center justify-center overflow-hidden rounded-2xl bg-slate-800 shadow-2xl"
@@ -274,10 +79,8 @@ export function VideoArea({ cameraOn, micOn, phase, basePose }: VideoAreaProps) 
         </AnimatePresence>
       </motion.div>
 
-      {/* 아랫부분: 실시간 수치 및 백엔드 전송 로그 영역 */}
       {phase === 'answering' && (
         <div className="grid grid-cols-2 gap-4 rounded-2xl border border-slate-700 bg-slate-900 p-4 text-white shadow-xl">
-          {/* 좌측 패널 (변경 없음) */}
           <div className="flex flex-col gap-2 rounded-xl bg-slate-800 p-3">
             <h3 className="flex items-center gap-2 text-sm font-semibold text-slate-300">
               <Activity className="h-4 w-4 text-emerald-400" />
@@ -311,7 +114,6 @@ export function VideoArea({ cameraOn, micOn, phase, basePose }: VideoAreaProps) 
             </div>
           </div>
 
-          {/* 우측 패널: 이벤트 기반 로그 */}
           <div className="flex flex-col gap-2 rounded-xl bg-slate-800 p-3">
             <h3 className="flex items-center gap-2 text-sm font-semibold text-slate-300">
               <AlertTriangle className="h-4 w-4 text-orange-400" />
